+++ src/tree/nodes.ts
@@ -1,9 +1,0 @@
-import * as vscode from 'vscode';
-
-interface TreeWithContext extends vscode.TreeItem {
-  readonly context: Record<string, any>;
-}
-
-export interface ReactVSCTreeNode {
-  readonly value: TreeWithContext;
-}