+++ src/treeReactApp/pages/index.tsx
@@ -1,3 +1,0 @@
-export * from './History';
-export * from './MyNotes';
-export * from './TeamNotes';